Added two more example interface sketches
Added example sketches for Teensy usbMIDI and Moppy2 serial. Have not verified Moppy2 serial yet as I'm having issues running the application. Relates to #1 and #9.

Consolidated SerialDebug.cpp into the header file. Will later move SerialDebug.h into the Common folder.

Changed methods related to pitch bending to use int for passing bend values between method calls. Also updated MIDI_Handler pitch bend handler to use int. Relates to #32.

+++ examples/Interface Examples/3Buzzers_Hairless_MIDI/Hairless_MIDI_Serial.h
@@ -2,7 +2,7 @@
 	#define Hairless_MIDI_Serial_h
 
 	#include <Arduino.h>
-	#include <MIDI_Device_Controller.h>
+	#include <MIDI_Handler\MIDI_Handler.h>
 	
 	class Hairless_MIDI_Serial
 	{
@@ -34,7 +34,7 @@
 			inline void setNoteOnHandle(void (*noteOn)(uint8_t, uint8_t, uint8_t)) { _midiHandler.setNoteOnHandle(noteOn); };
 			inline void setNoteOffHandle(void (*noteOff)(uint8_t, uint8_t, uint8_t)) { _midiHandler.setNoteOffHandle(noteOff); };
       inline void setControlChangeHandle(void (*controlChange)(uint8_t, uint8_t, uint8_t)) { _midiHandler.setControlChangeHandle(controlChange); };
-			inline void setPitchBendHandle(void (*pitchBend)(uint8_t, uint16_t)) { _midiHandler.setPitchBendHandle(pitchBend); };
+			inline void setPitchBendHandle(void (*pitchBend)(uint8_t, int)) { _midiHandler.setPitchBendHandle(pitchBend); };
 	};
 
 	//Defines a global instance of our class for users to consume
